Refactor: 용병 신청과 용병 컨트롤러 분리
